+++ src/sound/snd_cs423x.c
@@ -626,7 +626,7 @@ cs423x_ctxswitch_write(uint16_t addr, UNUSED(uint8_t val), void *priv)
 }
 
 static void
-cs423x_get_buffer(int32_t *buffer, int len, void *priv)
+cs423x_get_buffer(int32_t *buffer, uint16_t len, void *priv)
 {
     cs423x_t       *dev = (cs423x_t *) priv;
 
@@ -635,7 +635,7 @@ cs423x_get_buffer(int32_t *buffer, int len, void *priv)
 
     /* Don't output anything if the analog section or DAC is powered down. */
     if (!(dev->regs[2] & 0xb4) && !(dev->indirect_regs[9] & 0x04)) {
-        for (int c = 0; c < len * 2; c += 2) {
+        for (uint16_t c = 0; c < len * 2; c += 2) {
             buffer[c] += dev->ad1848.buffer[c] / 2;
             buffer[c + 1] += dev->ad1848.buffer[c + 1] / 2;
         }
@@ -645,7 +645,7 @@ cs423x_get_buffer(int32_t *buffer, int len, void *priv)
 }
 
 static void
-cs423x_get_music_buffer(int32_t *buffer, int len, void *priv)
+cs423x_get_music_buffer(int32_t *buffer, uint16_t len, void *priv)
 {
     cs423x_t *dev = (cs423x_t *) priv;
 
@@ -656,7 +656,7 @@ cs423x_get_music_buffer(int32_t *buffer, int len, void *priv)
         /* Don't output anything if the analog section, DAC (DAC2 instead on CS4235+) or FM synth is powered down. */
         uint8_t bpd_mask = (dev->type >= CRYSTAL_CS4235) ? 0xb1 : 0xb5;
         if (!(dev->regs[2] & bpd_mask) && !(dev->indirect_regs[9] & 0x06)) {
-            for (int c = 0; c < len * 2; c += 2) {
+            for (uint16_t c = 0; c < len * 2; c += 2) {
                 buffer[c] += (opl_buf[c] * dev->ad1848.fm_vol_l) >> 16;
                 buffer[c + 1] += (opl_buf[c + 1] * dev->ad1848.fm_vol_r) >> 16;
             }
