+++ src/sound/snd_wss.c
@@ -75,26 +75,26 @@ wss_write(UNUSED(uint16_t addr), uint8_t val, void *priv)
 }
 
 static void
-wss_get_buffer(int32_t *buffer, int len, void *priv)
+wss_get_buffer(int32_t *buffer, uint16_t len, void *priv)
 {
     wss_t *wss = (wss_t *) priv;
 
     ad1848_update(&wss->ad1848);
-    for (int c = 0; c < len * 2; c++)
+    for (uint16_t c = 0; c < len * 2; c++)
         buffer[c] += wss->ad1848.buffer[c] / 2;
 
     wss->ad1848.pos = 0;
 }
 
 static void
-wss_get_music_buffer(int32_t *buffer, int len, void *priv)
+wss_get_music_buffer(int32_t *buffer, uint16_t len, void *priv)
 {
     wss_t *wss = (wss_t *) priv;
     const int32_t *opl_buf = NULL;
 
     opl_buf = wss->opl.update(wss->opl.priv);
 
-    for (int c = 0; c < len * 2; c++) {
+    for (uint16_t c = 0; c < len * 2; c++) {
         if (opl_buf)
             buffer[c] += opl_buf[c];
     }
